requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
openai-whisper==20230314
spacy==3.5.0
textstat==0.7.3
language-tool-python==2.7.1
numpy==1.24.2
google-generativeai==0.3.1

